pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reflex-lasso-select"
version = "0.0.3"
description = "Reflex custom component wrapping  react-lasso-select"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [{ name = "Tim Child", email = "timjchild@gmail.com" }]
keywords = ["reflex", "reflex-custom-components", "lasso-select"]

dependencies = ["reflex>=0.7.1"]

classifiers = ["Development Status :: 4 - Beta"]

[project.urls]
homepage = "https://github.com/TimChild/reflex-lasso-select"
source = "https://github.com/TimChild/reflex-lasso-select"

[project.optional-dependencies]
dev = ["build", "twine"]



[tool.pytest.ini_options]
# addopts = "--headed"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.setuptools.packages.find]
where = ["custom_components"]



[dependency-groups]
dev = [
    "coverage>=7.6.12",
    "pre-commit>=4.1.0",
    "pyright>=1.1.396",
    "pytest>=8.3.5",
    "pytest-playwright>=0.7.0",
    "ruff>=0.9.10",
]
